requirements.txt
aiohttp==3.9.3
discord.py==2.4.0
loguru==0.7.2
matplotlib==3.10.0
pytz==2024.2
